requirements.txt
ccxt
pandas
numpy
schedule
python-telegram-bot>=20.0,<23.0
pytz
matplotlib
